tree-wide: remove unnecessary casts
Fixes clippy lint in Rust 1.68

BUG=b:276487055
TEST=tools/clippy # with rust 1.68

Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4391107
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>

+++ common/p9/src/server/tests.rs
@@ -976,7 +976,7 @@ fn setup_simple_lock(flags: u32) -> Server {
         ROOT_FID,
         filename,
         ROOT_FID + 1,
-        flags as u32,
+        flags,
     )
     .expect("failed to open file");
