+++ disk/src/qcow/mod.rs
@@ -639,7 +639,7 @@ impl QcowFile {
 
         // Set the refcount for each refcount table cluster.
         let cluster_size = 0x01u64 << qcow.header.cluster_bits;
-        let refcount_table_base = qcow.header.refcount_table_offset as u64;
+        let refcount_table_base = qcow.header.refcount_table_offset;
         let end_cluster_addr =
             refcount_table_base + u64::from(qcow.header.refcount_table_clusters) * cluster_size;
 
@@ -985,7 +985,7 @@ impl QcowFile {
     // Gets the offset of the given guest address in the host file. If L1, L2, or data clusters have
     // yet to be allocated, return None.
     fn file_offset_read(&mut self, address: u64) -> std::io::Result<Option<u64>> {
-        if address >= self.virtual_size() as u64 {
+        if address >= self.virtual_size() {
             return Err(std::io::Error::from_raw_os_error(EINVAL));
         }
 
@@ -1028,7 +1028,7 @@ impl QcowFile {
     // Gets the offset of the given guest address in the host file. If L1, L2, or data clusters need
     // to be allocated, they will be.
     fn file_offset_write(&mut self, address: u64) -> std::io::Result<u64> {
-        if address >= self.virtual_size() as u64 {
+        if address >= self.virtual_size() {
             return Err(std::io::Error::from_raw_os_error(EINVAL));
         }
 
@@ -1163,7 +1163,7 @@ impl QcowFile {
     // Deallocate the storage for the cluster starting at `address`.
     // Any future reads of this cluster will return all zeroes (or the backing file, if in use).
     fn deallocate_cluster(&mut self, address: u64) -> std::io::Result<()> {
-        if address >= self.virtual_size() as u64 {
+        if address >= self.virtual_size() {
             return Err(std::io::Error::from_raw_os_error(EINVAL));
         }
