+++ kvm/tests/read_only_memory.rs
@@ -44,7 +44,7 @@ fn test_run() {
     vcpu.set_sregs(&vcpu_sregs).expect("set sregs failed");
 
     let mut vcpu_regs: kvm_regs = unsafe { std::mem::zeroed() };
-    vcpu_regs.rip = load_addr.offset() as u64;
+    vcpu_regs.rip = load_addr.offset();
     vcpu_regs.rflags = 2;
     vcpu_regs.rax = 0x66;
     vcpu_regs.rbx = 0;
